@@ -35,65 +36,67 @@
   "private": true,
   "devDependencies": {
     "@types/classnames": "0.0.32",
-    "@types/electron": "^1.4.30",
+    "@types/electron": "^1.4.32",
     "@types/electron-devtools-installer": "^2.0.2",
-    "@types/mocha": "^2.2.35",
-    "@types/msgpack-lite": "^0.1.1",
-    "@types/node": "^6.0.55",
+    "@types/mocha": "^2.2.39",
+    "@types/msgpack-lite": "^0.1.3",
+    "@types/node": "^7.0.5",
     "@types/power-assert": "^1.4.29",
-    "@types/react": "^0.14.55",
-    "@types/react-dom": "^0.14.20",
+    "@types/react": "^15.0.10",
+    "@types/react-dom": "^0.14.23",
     "autogypi": "^0.2.2",
-    "babel-core": "^6.21.0",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-espower": "^2.3.1",
+    "babel-core": "^6.23.1",
+    "babel-loader": "^6.3.2",
+    "babel-plugin-espower": "^2.3.2",
     "babel-plugin-glslify": "^2.0.0",
-    "cross-env": "^3.1.3",
+    "cross-env": "^3.1.4",
     "css-loader": "^0.26.1",
-    "electron": "^1.4.13",
-    "electron-devtools-installer": "^2.0.1",
-    "electron-packager": "^8.4.0",
-    "electron-rebuild": "^1.4.0",
-    "file-loader": "^0.9.0",
+    "electron": "^1.6.0",
+    "electron-devtools-installer": "^2.1.0",
+    "electron-packager": "^8.5.1",
+    "electron-rebuild": "^1.5.7",
+    "file-loader": "^0.10.0",
     "json-loader": "^0.5.4",
     "mocha": "^3.2.0",
-    "mocha-loader": "github:seanchas116/mocha-loader#electron-renderer",
-    "nbind": "^0.3.8",
+    "mocha-loader": "^1.1.0",
+    "nbind": "^0.3.9",
     "node-gyp": "^3.5.0",
-    "npm-run-all": "^3.1.2",
+    "npm-run-all": "^4.0.1",
     "postcss-cssnext": "^2.9.0",
-    "postcss-import": "^9.0.0",
-    "postcss-loader": "^1.2.1",
+    "postcss-import": "^9.1.0",
+    "postcss-loader": "^1.3.0",
     "postcss-url": "^5.1.2",
     "power-assert": "^1.4.2",
-    "shelljs": "^0.7.5",
+    "resolve": "^1.2.0",
+    "shelljs": "^0.7.6",
     "stack-source-map": "^1.0.6",
     "style-loader": "^0.13.1",
-    "ts-loader": "^1.3.3",
-    "ts-node": "^2.0.0",
-    "tslint": "^4.2.0",
-    "typescript": "^2.1.4",
+    "ts-loader": "^2.0.0",
+    "ts-node": "^2.1.0",
+    "tslint": "^4.4.2",
+    "typescript": "^2.1.6",
     "url-loader": "^0.5.7",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2",
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.3.0",
     "webpack-fail-plugin": "^1.0.5"
   },
   "dependencies": {
     "classnames": "^2.2.5",
     "decamelize": "^1.2.0",
     "deep-assign": "^2.0.0",
+    "deep-equal": "^1.0.1",
     "glslify": "^6.0.1",
+    "keyboard-layout": "^2.0.11",
     "minimist": "^1.2.0",
-    "mobx": "^2.7.0",
-    "mobx-react": "^4.0.4",
+    "mobx": "^3.1.0",
+    "mobx-react": "^4.1.0",
     "msgpack-lite": "^0.1.26",
     "paintgl": "github:sketchglass/paintgl#64926a07ab5eadbc97845ec417e1e2acdfc35714",
     "paintvec": "github:sketchglass/paintvec#2c5deb2c4cea6e2b22e43cb5bf8b96115dc97455",
     "promise-semaphore": "^0.2.8",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-draggable-tree": "sketchglass/react-draggable-tree#9177b1b6de7f5bde9a8e1fe63d25bea0884ffde7",
-    "receive-tablet-event": "github:seanchas116/receive-tablet-event#8147b9191c67fb6c818ff9c6eab28e0e6fb699b7",
-    "rxjs": "^5.0.2"
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-draggable-tree": "sketchglass/react-draggable-tree#398cbfcb0c58088b865835faca1f6306603ac754",
+    "receive-tablet-event": "github:seanchas116/receive-tablet-event#8147b9191c67fb6c818ff9c6eab28e0e6fb699b7"
   }
 }
