@@ -1,6 +1,6 @@
 import {PictureAction} from "./Action"
 import ActionIDs from "./ActionIDs"
-import {addAction} from "../state/ActionRegistry"
+import {addAction} from "../app/ActionRegistry"
 import {dialogLauncher} from "../views/dialogs/DialogLauncher"
 import {FlipPictureCommand, Rotate90PictureCommand, Rotate180PictureCommand, ChangePictureResolutionCommand} from "../commands/PictureCommand"
 
