@@ -1,8 +1,8 @@
 import {remote} from "electron"
 import Action, {PictureAction} from "./Action"
 import ActionIDs from "./ActionIDs"
-import {addAction} from "../state/ActionRegistry"
-import {appState} from "../state/AppState"
+import {addAction} from "../app/ActionRegistry"
+import {appState} from "../app/AppState"
 
 @addAction
 export class ViewReloadAction extends Action {
@@ -20,7 +20,7 @@ export class ViewToggleDevToolsAction extends Action {
   title = "Toggle Developer Tools"
   enabled = true
   run() {
-    remote.getCurrentWebContents().toggleDevTools()
+    remote.BrowserWindow.getFocusedWindow().webContents.toggleDevTools()
   }
 }
 
