@@ -1,25 +1,24 @@
 import {action} from "mobx"
 import * as React from "react"
 import {observer} from "mobx-react"
-import {remote} from "electron"
 import * as classNames from "classnames"
+import {remote} from "electron"
+const {Menu} = remote
+
+import KeyInput from "../../lib/KeyInput"
 
 import SVGIcon from "./components/SVGIcon"
 
-import {appState} from "../state/AppState"
+import {toolManager} from "../app/ToolManager"
 
 import Tool from "../tools/Tool"
 import ToolIDs from "../tools/ToolIDs"
-import BaseBrushTool from "../tools/BaseBrushTool"
-import BrushTool from "../tools/BrushTool"
-import WatercolorTool from "../tools/WatercolorTool"
 
-const {Menu} = remote
+import {dialogLauncher} from "../views/dialogs/DialogLauncher"
 
 const toolToIcon = (tool: Tool) => {
   const map = {
-    [ToolIDs.watercolor]: "paint-brush",
-    [ToolIDs.brush]: "pen",
+    [ToolIDs.brush]: "paint-brush",
     [ToolIDs.pan]: "move",
     [ToolIDs.rotate]: "rotate",
     [ToolIDs.zoom]: "search",
@@ -39,52 +38,42 @@ export default
 class ToolSelection extends React.Component<{hidden: boolean}, {}> {
   render() {
     const {hidden} = this.props
-    const {tools, currentTool} = appState
-    const nonBrushTools = tools.filter(tool => !(tool instanceof BaseBrushTool))
-    const brushTools = tools.filter(tool => tool instanceof BaseBrushTool)
+    const {tools, currentTool} = toolManager
     return (
       <div className="ToolSelection" hidden={hidden}>{
-          nonBrushTools.map((tool, i) => {
-            const selected = tool === currentTool
-            const className = classNames("ToolSelection_button", {"ToolSelection_button-selected": selected})
-            const onClick = () => this.onChange(tool)
-            return <button key={i} onClick={onClick} className={className}>{toolToIcon(tool)}</button>
-          })
-        }{
-          brushTools.map((tool, i) => {
-            const onContextMenu = (e: React.MouseEvent<Element>) => this.onContextMenu(tool, e)
-            const selected = tool === currentTool
-            const className = classNames("ToolSelection_button", {"ToolSelection_button-selected": selected})
-            const onClick = () => this.onChange(tool)
-            return <button key={i} onContextMenu={onContextMenu} onClick={onClick} className={className}>{toolToIcon(tool)}</button>
-          })
-        }
+        tools.map((tool, i) => {
+          const selected = tool === currentTool
+          const className = classNames("ToolSelection_button", {"ToolSelection_button-selected": selected})
+          const onClick = () => this.onChange(tool)
+          const onContextMenu = (e: React.MouseEvent<HTMLElement>) => this.onContextMenu(e, tool)
+          return <button key={i} onClick={onClick} onContextMenu={onContextMenu} className={className}>{toolToIcon(tool)}</button>
+        })
+      }
       </div>
     )
   }
   private onChange = action((tool: Tool) => {
-    appState.currentTool = tool
+    toolManager.currentTool = tool
   })
-  private onContextMenu = action((selectedTool: Tool, e: React.MouseEvent<Element>) => {
-    e.preventDefault()
-    const removeTool = action(() => {
-      const index = appState.tools.indexOf(selectedTool)
-      appState.tools.splice(index, 1)
-    })
-    const appendTool = action((item: Tool) => {
-      return () => {
-        const index = appState.tools.indexOf(selectedTool) + 1
-        appState.tools.splice(index, 0, item)
+  private onContextMenu = action((e: React.MouseEvent<HTMLElement>, tool: Tool) => {
+    toolManager.currentTool = tool
+    const {clientX, clientY} = e
+    setTimeout(() => {
+      const selectShortcuts = async () => {
+        const result = await dialogLauncher.openToolShortcutsDialog({
+          toggle: tool.toggleShortcut && tool.toggleShortcut.toData(),
+          temp: tool.tempShortcut && tool.tempShortcut.toData(),
+        })
+        if (result) {
+          const {toggle, temp} = result
+          tool.toggleShortcut  = toggle && KeyInput.fromData(toggle)
+          tool.tempShortcut = temp && KeyInput.fromData(temp)
+        }
       }
-    })
-    const menuTemplate = [
-      {label: '追加', submenu: [
-        {label: "BrushTool", click: appendTool(new BrushTool())},
-        {label: "WatercolorTool", click: appendTool(new WatercolorTool())}
-      ]},
-      {label: '削除', click: removeTool}
-    ]
-    const menu = Menu.buildFromTemplate(menuTemplate)
-    menu.popup(remote.getCurrentWindow())
+      const menu = Menu.buildFromTemplate([
+        {label: "Shortcuts...", click: selectShortcuts},
+      ])
+      menu.popup(remote.getCurrentWindow(), clientX, clientY)
+    }, 50)
   })
 }
