@@ -1,5 +1,7 @@
 import {PictureDimension} from "../../models/Picture"
 import {ipcRenderer} from "electron"
+import IPCChannels from "../../../common/IPCChannels"
+import {ToolShortcutsDialogData} from "./ToolShortcutsDialog"
 
 export default
 class DialogLauncher {
@@ -12,16 +14,20 @@ class DialogLauncher {
     return this.open<PictureDimension, PictureDimension>("resolutionChange", init)
   }
 
+  openToolShortcutsDialog(init: ToolShortcutsDialogData) {
+    return this.open<ToolShortcutsDialogData, ToolShortcutsDialogData>("toolShortcuts", init)
+  }
+
   async open<TResult, TParam>(name: string, param: TParam): Promise<TResult|undefined> {
     let callback: any
     const result = await new Promise<TResult|undefined>((resolve, reject) => {
       callback = (e: Electron.IpcRendererEvent, result: TResult|undefined) => {
         resolve(result)
       }
-      ipcRenderer.on("dialogDone", callback)
-      ipcRenderer.send("dialogOpen", name, param)
+      ipcRenderer.on(IPCChannels.dialogDone, callback)
+      ipcRenderer.send(IPCChannels.dialogOpen, name, param)
     })
-    ipcRenderer.removeListener("dialogDone", callback)
+    ipcRenderer.removeListener(IPCChannels.dialogDone, callback)
     return result
   }
 }
