@@ -2,7 +2,7 @@ import React = require("react")
 import {observer} from "mobx-react"
 import {PictureDimension} from "../../models/Picture"
 import DimensionSelect from "../DimensionSelect"
-import DimensionSelectState from "../../state/DimensionSelectState"
+import DimensionSelectViewModel from "../../viewmodels/DimensionSelectViewModel"
 import DialogContainer from "./DialogContainer"
 
 interface NewPictureDialogProps {
@@ -13,12 +13,12 @@ interface NewPictureDialogProps {
 @observer
 export default
 class NewPictureDialog extends React.Component<NewPictureDialogProps, {}> {
-  private dimensionSelectState = new DimensionSelectState()
+  private dimensionSelectViewModel = new DimensionSelectViewModel()
 
   render() {
     return (
-      <DialogContainer okText="New" canOK={this.dimensionSelectState.isValid} onOK={this.onOK} onCancel={this.onCancel}>
-        <DimensionSelect state={this.dimensionSelectState} />
+      <DialogContainer title="New Picture" okText="New" canOK={this.dimensionSelectViewModel.isValid} onOK={this.onOK} onCancel={this.onCancel}>
+        <DimensionSelect viewModel={this.dimensionSelectViewModel} />
       </DialogContainer>
     )
   }
@@ -32,6 +32,6 @@ class NewPictureDialog extends React.Component<NewPictureDialogProps, {}> {
   }
 
   private onOK = () => {
-    this.props.onDone(this.dimensionSelectState.dimension)
+    this.props.onDone(this.dimensionSelectViewModel.dimension)
   }
 }
